requirements.txt
requests
beautifulsoup4
lxml
python-telegram-bot==20.7
apscheduler==3.10.4
